backend_fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "jskim1102",email = "js.kim.deepi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.42,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-auth (>=2.40.3,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.7"
isort = "^6.0.1"
flake8 = "^7.3.0"

